backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
onnxruntime
numpy
opencv-python
albumentations
Pillow
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
motor
pymongo
email-validator
aiohttp
rasterio
matplotlib
scipy
pytest
alle-ai-sdk
requests